+++ geometry/geometry_config.py
@@ -231,8 +231,8 @@ with ConfigRegistry.register_config("basic") as c:
 
     c.SplitCal = AttrDict(z=0)
     c.SplitCal.ZStart = 38.450 * u.m  # Relative start z of split cal to decay vessel centre
-    c.SplitCal.XMax = 480.*u.cm/2. #290.*u.cm  #half length
-    c.SplitCal.YMax = 720. * u.cm / 2. #510.*u.cm * c.Yheight / (10.*u.m)   #half length
+    c.SplitCal.XMax = 4 * u.m / 2  # half length
+    c.SplitCal.YMax = 6 * u.m / 2  # half length
     c.SplitCal.Empty = 0*u.cm
     c.SplitCal.BigGap = 100*u.cm
     c.SplitCal.ActiveECALThickness = 0.56*u.cm
@@ -256,8 +256,8 @@ with ConfigRegistry.register_config("basic") as c:
     c.SplitCal.NModulesInX = 2
     c.SplitCal.NModulesInY = 3
     c.SplitCal.NStripsPerModule = 50
-    c.SplitCal.StripHalfWidth = 2 * u.cm # c.SplitCal.XMax/(c.SplitCal.NStripsPerModule*c.SplitCal.NModulesInX)
-    c.SplitCal.StripHalfLength = 100 * u.cm # c.SplitCal.YMax/c.SplitCal.NModulesInY
+    c.SplitCal.StripHalfWidth = c.SplitCal.XMax / (c.SplitCal.NStripsPerModule * c.SplitCal.NModulesInX)
+    c.SplitCal.StripHalfLength = c.SplitCal.YMax / c.SplitCal.NModulesInY
     c.SplitCal.SplitCalThickness=(c.SplitCal.FilterECALThickness_first-c.SplitCal.FilterECALThickness)+(c.SplitCal.FilterECALThickness+c.SplitCal.ActiveECALThickness)*c.SplitCal.nECALSamplings+c.SplitCal.BigGap
 
     zecal = 38.450 * u.m  # Relative start z of ECAL to decay vessel centre